refactor (components): rename CommonPage.tsx to CrudHeader.tsx
